Added working example python bindings (finally)
After spending entire first part of the semester working on this, I have finally found a working method for this.
Currently, this is code from another repository (the official pybind11 repository) and I plan to expand it further.

The repository can be found here https://github.com/pybind/cmake_example

Some of the methods I have tried and have not worked can be found at the links provided

-https://realpython.com/python-bindings-overview/ (a method involving using pybind11 and invoke to build and test, this wouldn't
work for me because invoke did not want to work on my machine)

-https://www.blopig.com/blog/2021/03/c-python-bindings-in-5-minutes/ (a method not involving cmake, which is a big no for the longevity
of Glibby as it used cmake. However, I still tried this and it did not work as it never recognized pybind11.h as a module, nor the python version
of the import)

-https://docs.python.org/3/extending/extending.html (This method ended up being too confusing in the end to follow correctly)

-https://www.geeksforgeeks.org/how-to-call-c-c-from-python/ (This method was a red herring as it did not work with any non-trial
functions A.K.A., anything that didn't just print a line. This method also made me look back at other methods to see if I could possibly
expand upon the very basic "working" version this provided which ended up just being a useless timesink in the end)

-https://martinopilia.com/posts/2018/09/15/building-python-extension.html (This also did not work for similar reasons to above methods)

-https://pybind11.readthedocs.io/en/stable/basics.html (This did not work for me as cmake, and make did not want to work correctly on
my machine which sent me down a wild goosechase of installs and downloads that didn't work. However, this method did yield a link to the
official pybind11 repo that had exactly what I needed to make this work, but I didn't notice this until the 2nd or 3rd attempt to use this method)

-https://learn.microsoft.com/en-us/visualstudio/python/working-with-c-cpp-python-in-visual-studio?view=vs-2022 (This also did not work for similar
reasons to above)

-https://wiki.python.org/moin/IntegratingPythonWithOtherLanguages (This lead me to the method that eventually worked!)

I also remember one method that I tried (that involved a distinct folder structure, hence why I remember it) but I am not able to find a link to it. This method
also did not work because of not recognizing pybind11 or the python import (similar to above)

In addition, I also looked at numerous stack overflow threads along the way for the overarching problem (the bindings) as well as method specific problems
(like why make and its commands would not work for me, this was included in the "wild goosechase of installs and downloads" method)

However, a basic addition and subtraction works here. To call it, I used ubuntu which allowed it to work.