Point.h bindings compile, do not run yet
Bindings for actual functions are now underway and the bindings for Point.h are now finally compiling.
Team lead Wyatt has helped me these past weeks integrating the python bindings into the main build so that
I can access the files I need to make bindings too from within my python binding file.

However, a multitude of issues arose during this:

-I am still relatively new to CMake and getting pybind11 included correctly took a lot of work and a substancial
 amount of help from Wyatt (he was the one that got it to work in the end).

-Multiple different attempts were made to get them to compile and into the project which I wanted to initially avoid
 at first due to the python bindings being strictly optional for our library. However, having them not directly integrated
 into the main build is currently beyond my skill level at the moment.

Additionally, the bindings for Point.h were not easy to implement due to the functions within the file being template
functions with more than 1 argument. (For example template<typename T> is relatively easy to inplement the bindings for, and
I was able to actually produce this with an example from the pybind11 documentation which can be found here:
https://pybind11.readthedocs.io/en/stable/advanced/functions.html#binding-functions-with-template-parameters )

As seen in the above link, the way to perform said binding for:

	template <typename T>
	void set(T t);

would be:

	m.def("set", &set<your-type-here>)

However, this is not nearly as simple when the template has more than 1 parameter.

Bindings for ManhattanDistance were written like this:

	m.def("ManhattanDistance", glibby::ManhattanDistance<glibby::Point<type-T, numb-1-through-3>>)

	*NOTE*: the function template is - template<typename T, int N>

However, the binding line for m.def would not recognize the types going into the function ManhattanDistance. I temporarily
(as I really want to change this in the future) solved this with a proxy function defined here:

	template<typename T>
	T ManhattanDistance1D(const glibby::Point<T, 1>& p1, const glibby::Point<T, 1>& p2)
	{
		return glibby::ManhattanDistance(p1, p2);
	}

This solves the case of the template problem and allows the code to compile, however it is clunky and I wish to refine it in
the future.

This approach was also how EuclidianDistance was done as well.